Upgrade print/copy rights from int32 to int64 and fix SQL Server schema types
Widen UserRights.Print and .Copy from *int32 to *int64 to support rights
values beyond 2 billion. Update all database schemas (SQL Server, PostgreSQL,
MySQL, SQLite) to use bigint for rights_print and rights_copy columns. Fix
SQL Server content.length from int to bigint for files >2GB. Update config
and frontend to match the new int64 types.

+++ config/config.go
@@ -55,8 +55,8 @@ type FrontendServerInfo struct {
 	ServerInfo          `yaml:",inline"`
 	Directory           string `yaml:"directory,omitempty"`
 	ProviderUri         string `yaml:"provider_uri"`
-	RightPrint          int32  `yaml:"right_print"`
-	RightCopy           int32  `yaml:"right_copy"`
+	RightPrint          int64  `yaml:"right_print"`
+	RightCopy           int64  `yaml:"right_copy"`
 	MasterRepository    string `yaml:"master_repository"`
 	EncryptedRepository string `yaml:"encrypted_repository"`
 }
